Create organize routes component
